[Sharding] Move Synchronizer Test code, fix  namespace
